[Feat] 포지션 및 기술스택 기능 구현 (#4)
* :memo: Position & TechStack SQL 데이터 파일 생성

* :bulb: EmptyException 추가

* :sparkles: 포지션 조회 기능 추가

* :fire:

* :white_check_mark: 포지션 조회 테스트 완료

* :sparkles: 기술스택 조회 기능 추가

* :recycle: TechStackRequestDto -> TechStackDto로 변경

* :recycle: techStackId Long에서 Short로 변경

* :white_check_mark: TechStack 조회 기능 테스트 완료

* :recycle: techStack api 추가

* :recycle:

+++ src/main/java/org/example/develociraptor/global/idclass/UserTechStackId.java
@@ -16,5 +16,5 @@ public class UserTechStackId implements Serializable {
 
 	private Long userId;
 
-	private Long techStackId;
+	private Short techStackId;
 }
